backend/requirements.txt
absl-py==2.1.0
addict==2.4.0
antlr4-python3-runtime==4.13.1
attrs==23.2.0
basicsr==1.4.2
certifi==2023.7.22
charset-normalizer==3.3.2
clip==1.0
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
ffmpeg-python==0.2.0
filelock==3.13.1
filterpy==1.1.0
fonttools==4.49.0
fsspec==2024.2.0
ftfy==6.1.3
future==1.0.0
grpcio==1.62.0
idna==3.4
imageio==2.34.0
importlib-metadata==7.0.1
Jinja2==3.1.3
jsonmerge==1.9.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lazy_loader==0.3
lmdb==1.4.1
Markdown==3.5.2
MarkupSafe==2.1.5
matplotlib==3.8.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
opencv-python==4.9.0.80
optional-django==0.3.0
packaging==23.2
pillow==10.2.0
platformdirs==4.2.0
protobuf==4.25.3
pyparsing==3.1.1
python-dateutil==2.8.2
PyYAML==6.0.1
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rpds-py==0.18.0
scikit-image==0.22.0
scipy==1.12.0
six==1.16.0
sympy==1.12
tb-nightly==2.17.0a20240225
tensorboard-data-server==0.7.2
tifffile==2024.2.12
tomli==2.0.1
torch==2.2.1
torchvision==0.17.1
tqdm==4.66.2
typing_extensions==4.10.0
urllib3==2.0.7
wcwidth==0.2.13
webpack==6.0.0
Werkzeug==3.0.1
yapf==0.40.2
zipp==3.17.0
